refactor(web): extract shared diff and endpoint primitives
Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
