+++ apps/web/src/shared/types/endpoint.ts
@@ -1,0 +1,3 @@
+import type { ResourceCategory } from '@cloudblocks/schema';
+
+export type EndpointType = ResourceCategory | 'internet' | 'browser';
